Results Upload, Scripts to Directory
Uploaded the results of the ensemble modeling, and moved all scripts from main directory to a dedicated script directory. Updated modeling scripts to work with stsb.
