+++ Results/GradBoostEnsemble_save.csv
@@ -1,0 +1,10 @@
+,Task,Ensemble,DeBERTa,Electra,XLNet,Ensemble_f,DeBERTa_f,Electra_f,XLNet_f
+0,cola,0.6641392105403996,0.6208288813242873,0.6070286836001448,0.3999158683209528,999.0,999,999,999.0
+1,mnli,0.8777589134125636,0.8745797249108508,0.8174223127865512,0.8566479877738156,999.0,999,999,999.0
+2,mnli-mm,0.871864406779661,0.8723555736371034,0.8212978030919447,0.8578112286411717,999.0,999,999,999.0
+3,mrpc,0.8699186991869918,0.8651960784313726,0.8823529411764706,0.8921568627450981,0.9047619047619048,0.9029982363315697,0.9145907473309609,0.9225352112676057
+4,qnli,0.9206833435021354,0.9154310818231741,0.8890719384953323,0.8779059125022881,999.0,999,999,999.0
+5,qqp,0.9101327397147333,0.9063071976255256,0.9003215434083601,0.8737818451644819,0.8793980969240982,0.8749917497194905,0.8659972068896721,0.8333170014698678
+6,rte,0.7142857142857143,0.6678700361010831,0.6823104693140795,0.740072202166065,999.0,999,999,999.0
+7,sst2,0.9580152671755725,0.9346330275229358,0.9174311926605505,0.9403669724770642,999.0,999,999,999.0
+8,wnli,0.6363636363636364,0.43661971830985913,0.4647887323943662,0.5633802816901409,999.0,999,999,999.0
