+++ Results/Predictions/gradient_boost_fit1_wnli.csv
@@ -1,0 +1,146 @@
+1.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+0.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
+1.000000000000000000e+00
